Move DB-related function from testing package to inmanta.db.util
